Refactor editors into dedicated class
